refactor(router): move to plugins directory
- I also replaced @ imports with absolute paths
