+++ frontend/cypress/plugins/index.js
@@ -1,7 +1,7 @@
 /** @file custom handlers for cypress events */
 /// <reference types="cypress" />
 // ***********************************************************
-// This example plugins/index.js can be used to load plugins
+// This example plugins/router.js can be used to load plugins
 //
 // You can change the location of this file or turn off loading
 // the plugins file with the 'pluginsFile' configuration option.
