+++ frontend/src/main.js
@@ -4,7 +4,7 @@ import {createApp} from 'vue';
 import App from './App.vue';
 import store from './store/index';
 import VueAxios from 'vue-axios';
-import router from './router/index';
+import router from './plugins/router';
 import i18n from './i18n';
 import vuetify from './plugins/vuetify';
