+++ frontend/src/plugins/router.js
@@ -1,13 +1,13 @@
 /** @file router configuration */
 
 import {createRouter, createWebHistory} from 'vue-router';
-import Register from '@/views/Register.vue';
-import Home from '@/views/Home.vue';
-import Login from '@/views/Login.vue';
-import NotFound from '@/views/NotFound.vue';
-import CreateColor from '@/views/CreateColor.vue';
-import CreateGym from '@/views/CreateGym.vue';
-import GymMapView from '@/views/GymMap.vue';
+import Register from '../views/Register.vue';
+import Home from '../views/Home.vue';
+import Login from '../views/Login.vue';
+import NotFound from '../views/NotFound.vue';
+import CreateColor from '../views/CreateColor.vue';
+import CreateGym from '../views/CreateGym.vue';
+import GymMapView from '../views/GymMap.vue';
 
 export default createRouter({
   history: createWebHistory(),
